resolve errors and duplicate classes after previous merging by other colleague
