+++ src/main/java/faang/school/postservice/config/redis/RedisConfiguration.java
@@ -2,19 +2,37 @@ package faang.school.postservice.config.redis;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.redis.connection.RedisConnectionFactory;
+import org.springframework.data.redis.connection.RedisStandaloneConfiguration;
 import org.springframework.data.redis.connection.jedis.JedisConnectionFactory;
 import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.data.redis.listener.ChannelTopic;
 import org.springframework.data.redis.serializer.Jackson2JsonRedisSerializer;
 
 @Configuration
 public class RedisConfiguration {
 
+    @Value("${spring.data.redis.channels.calculations_channel.name}")
+    private String topic;
+
+    @Value("${spring.data.redis.host}")
+    private String host;
+    @Value("${spring.data.redis.port}")
+    private int port;
+
     @Bean
     public JedisConnectionFactory jedisConnectionFactory() {
-        return new JedisConnectionFactory();
+        RedisStandaloneConfiguration configuration =
+                new RedisStandaloneConfiguration(host, port);
+        return new JedisConnectionFactory(configuration);
+    }
+
+    @Bean
+    public ChannelTopic calculations_channelTopic() {
+        return new ChannelTopic(topic);
     }
 
     @Bean
